+++ src/scmdata/errors.py
@@ -10,6 +10,7 @@ class NonUniqueMetadataError(ValueError):
     """
 
     def __init__(self, meta):
+        self.meta = meta
         # format table to show the metadata clash
         dup = meta.astype(str).groupby(meta.columns.tolist(), as_index=False).size()
         if isinstance(dup, pd.Series):
@@ -18,6 +19,7 @@ class NonUniqueMetadataError(ValueError):
             dup = dup.to_frame().reset_index()
         else:
             dup = dup.rename(columns={"size": "repeats"})
+
         dup = dup[dup.repeats > 1]
         msg = (
             "Duplicate metadata (numbers show how many times the given "
@@ -33,7 +35,7 @@ class MissingRequiredColumnError(ValueError):
     """
 
     def __init__(self, columns):
-        # self.columns = columns
+        self.columns = columns
         msg = "Missing required columns `{}`!".format(columns)
 
         super().__init__(msg)
@@ -43,7 +45,9 @@ class DuplicateTimesError(ValueError):
     """
     Error raised when times are duplicated
     """
+
     def __init__(self, time_index):
+        self.time_index = time_index
         dup = time_index.value_counts()
         dup = dup[dup > 1]